--- src/main/kotlin/com/aeyoung/together/domain/study/repository/StudyNoticeRepository.kt
@@ -1,9 +1,0 @@
-package com.aeyoung.together.domain.study.repository
-
-import com.aeyoung.together.domain.study.StudyNotice
-import com.aeyoung.together.domain.study.StudyRoom
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-interface StudyNoticeRepository : JpaRepository<StudyNotice, Long> {
-}
