--- src/main/kotlin/com/aeyoung/together/domain/study/service/WriteNoticeCommentService.kt
@@ -1,8 +1,0 @@
-package com.aeyoung.together.domain.study.service
-
-import com.aeyoung.together.domain.study.presentation.dto.req.WriteNoticeCommentReqDto
-
-interface WriteNoticeCommentService {
-
-    fun execute(writeNoticeCommentReqDto: WriteNoticeCommentReqDto,noticeId:Long)
-}
