--- src/main/kotlin/com/aeyoung/together/domain/study/service/WriteNoticeService.kt
@@ -1,8 +1,0 @@
-package com.aeyoung.together.domain.study.service
-
-import com.aeyoung.together.domain.study.presentation.dto.req.WriteNoticeReqDto
-
-interface WriteNoticeService {
-
-    fun execute(writeNoticeReqDto: WriteNoticeReqDto, studyId: Long)
-}
