+++ src/main/kotlin/com/aeyoung/together/domain/studyRoom/StudyNotice.kt
@@ -1,6 +1,7 @@
-package com.aeyoung.together.domain.study
+package com.aeyoung.together.domain.studyRoom
 
 import com.aeyoung.together.domain.member.Member
+import com.aeyoung.together.domain.study.StudyRoom
 import com.aeyoung.together.global.entity.BaseIdEntity
 import jakarta.persistence.Entity
 import jakarta.persistence.JoinColumn
