+++ src/main/kotlin/com/aeyoung/together/domain/studyRoom/presentation/dto/req/WriteNoticeCommentReqDto.kt
@@ -1,8 +1,8 @@
-package com.aeyoung.together.domain.study.presentation.dto.req
+package com.aeyoung.together.domain.studyRoom.presentation.dto.req
 
 import com.aeyoung.together.domain.member.Member
-import com.aeyoung.together.domain.study.StudyNotice
-import com.aeyoung.together.domain.study.StudyNoticeComment
+import com.aeyoung.together.domain.studyRoom.StudyNotice
+import com.aeyoung.together.domain.studyRoom.StudyNoticeComment
 import jakarta.validation.constraints.NotBlank
 
 class WriteNoticeCommentReqDto(
