+++ src/main/kotlin/com/aeyoung/together/domain/studyRoom/presentation/dto/req/WriteNoticeReqDto.kt
@@ -1,8 +1,8 @@
-package com.aeyoung.together.domain.study.presentation.dto.req
+package com.aeyoung.together.domain.studyRoom.presentation.dto.req
 
 import com.aeyoung.together.domain.member.Member
 //import com.aeyoung.together.domain.study.Comment
-import com.aeyoung.together.domain.study.StudyNotice
+import com.aeyoung.together.domain.studyRoom.StudyNotice
 import com.aeyoung.together.domain.study.StudyRoom
 import jakarta.validation.constraints.NotBlank
