+++ src/main/kotlin/com/aeyoung/together/domain/studyRoom/repository/StudyNoticeCommentRepository.kt
@@ -1,6 +1,6 @@
-package com.aeyoung.together.domain.study.repository
+package com.aeyoung.together.domain.studyRoom.repository
 
-import com.aeyoung.together.domain.study.StudyNoticeComment
+import com.aeyoung.together.domain.studyRoom.StudyNoticeComment
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface StudyNoticeCommentRepository : JpaRepository<StudyNoticeComment, Long> {
