+++ src/main/kotlin/com/aeyoung/together/domain/studyRoom/repository/StudyNoticeRepository.kt
@@ -1,0 +1,7 @@
+package com.aeyoung.together.domain.studyRoom.repository
+
+import com.aeyoung.together.domain.studyRoom.StudyNotice
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface StudyNoticeRepository : JpaRepository<StudyNotice, Long> {
+}
