+++ src/main/kotlin/com/aeyoung/together/domain/studyRoom/service/impl/WriteNoticeServiceImpl.kt
@@ -1,14 +1,12 @@
-package com.aeyoung.together.domain.study.service.impl
+package com.aeyoung.together.domain.studyRoom.service.impl
 
-import com.aeyoung.together.domain.study.StudyNotice
 import com.aeyoung.together.domain.study.exception.StudyNotFoundException
-import com.aeyoung.together.domain.study.presentation.dto.req.WriteNoticeReqDto
-import com.aeyoung.together.domain.study.repository.StudyNoticeRepository
+import com.aeyoung.together.domain.studyRoom.presentation.dto.req.WriteNoticeReqDto
+import com.aeyoung.together.domain.studyRoom.repository.StudyNoticeRepository
 import com.aeyoung.together.domain.study.repository.StudyRoomRepository
-import com.aeyoung.together.domain.study.service.WriteNoticeService
+import com.aeyoung.together.domain.studyRoom.service.WriteNoticeService
 import com.aeyoung.together.global.util.MemberUtil
 import org.springframework.stereotype.Service
-import java.lang.RuntimeException
 
 @Service
 class WriteNoticeServiceImpl(
